% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffd_bin_packing.R
\name{ffd_bin_packing}
\alias{ffd_bin_packing}
\title{First-fit-decreasing bin packing algorithm}
\usage{
ffd_bin_packing(games, storage)
}
\arguments{
\item{games}{a vector of games' sizes}

\item{storage}{a vector of storage facilities' sizes}
}
\value{
the matrix of storaged games
}
\description{
Optimize the storage of games
}
