% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_opti.R
\name{naive_storage_opti}
\alias{naive_storage_opti}
\title{naive storage optimisation algorithm}
\usage{
naive_storage_opti(storage, games)
}
\arguments{
\item{storage}{a vector of storage facilities' sizes}

\item{games}{a vector of games' sizes}
}
\value{
the matrix of storaged games
}
\description{
Optimize the storage of games
}
