% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/ffd_bin_packing.R
\name{ffd_bin_packing}
\alias{ffd_bin_packing}
\title{First-fit-decreasing bin packing algorithm}
\usage{
ffd_bin_packing(games, storage)

ffd_bin_packing(games, storage)
}
\arguments{
\item{games}{a vector of games' sizes}

\item{storage}{a vector of storage facilities' sizes}

\item{items}{A vector containing the sizes of items to be packed.}

\item{bin_size}{The maximum capacity of each bin.}
}
\value{
A vector of vectors representing the bins, where each inner vector
        contains the sizes of items packed into a single bin.

the matrix of storaged games
}
\description{
Optimize the storage of games
}
